# docs/swagger/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: |-
    YunShu CMDB is an operations CMDB console built with Gin, MySQL, Redis, SMTP mail delivery, Casbin, Cobra and slog.
    Request an email verification code first, then login with email code or register a new account by email. Username/password login remains available as an admin fallback.
  title: YunShu CMDB API
  version: "1.0"
paths:
...
  /api/v1/pods/exec/ws:
    get:
      description: WebSocket shell into a pod. Requires cluster_id, namespace, name
        query params. Obtain ticket via POST /api/v1/auth/ws-ticket first.
      parameters:
      - description: Cluster ID
        in: query
        name: cluster_id
        required: true
        type: integer
      - description: Pod namespace
        in: query
        name: namespace
        required: true
        type: string
      - description: Pod name
        in: query
        name: name
        required: true
        type: string
      - description: Container name
        in: query
        name: container
        type: string
      - description: One-time WebSocket ticket
        in: query
        name: ticket
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Interactive pod exec (WebSocket)
      tags:
      - Pods
...
  /api/v1/projects/{id}/servers/{serverId}/terminal/ws:
    get:
      description: WebSocket upgrade for project server SSH. Obtain ticket via POST
        /api/v1/auth/ws-ticket first, then pass ticket= in query string.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Server ID
        in: path
        name: serverId
        required: true
        type: integer
      - description: One-time WebSocket ticket
        in: query
        name: ticket
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Interactive SSH terminal (WebSocket)
      tags:
      - CMDB
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the JWT token with the `Bearer ` prefix, for example: Bearer
      eyJhbGciOi...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"